// src/models/WorkOrderCustomer.ts
export interface IWorkOrderCustomer {
    orderId: string
    customer: string
    partnerRole: string,
    customerName: string
    address: string
    phone: string
    country: string
    paymentType: string
    locationCode: string
    tradeName: string
    customerClass: string
    longitude: number
    latitude: number
    monMorFrom: string
    monMorTo: string
    monEventFrom: string
    monEventTo: string
    tueMorFrom: string
    tueMorTo: string
    tueEventFrom: string
    tueEventTo: string
    wedMorFrom: string
    wedMorTo: string
    wedEventFrom: string
    wedEventTo: string
    thuMorFrom: string
    thuMorTo: string
    thuEventFrom: string
    thuEventTo: string
    friMorFrom: string
    friMorTo: string
    friEventFrom: string
    friEventTo: string
    satMorFrom: string
    satMorTo: string
    satEventFrom: string
    satEventTo: string
    sunMorFrom: string
    sunMorTo: string
    sunEventFrom: string
    sunEventTo: string
}